Fix #16 by adding form[action] attributes everywhere
Apparently Safari on iOS will not show a submit key on the keyboard that comes up when focussing into the details textarea when adding timeboxes (and possibly the other forms, haven't tested but assuming them broken just in case) _unless_ the form has an action attribute. This commit adds `action=""` to all three forms currently in the app.