Another attempt at fixing #16 by using newlines to submit the form
It does not appear possible to get a "Go" button to show up when editing a text area (which does make sense). We're only abusing a textarea because there is no way to make a text input wrap its content. This new attempt checks for a newline at the end of the input value and submits the form if it detects one. This should allow iOS users to use the return key on the presented keyboard to add a timebox.